登陆认证修改：统一url等
