+++ src/main/java/com/vanke/tydirium/constants/CommonConstants.java
@@ -9,8 +9,9 @@ package com.vanke.tydirium.constants;
  * @date: 2017年8月23日 下午3:29:12
  */
 public class CommonConstants {
-	/**
-	 * 分页page对象标识
-	 */
+	// 分页page对象标识
 	public static final String PAGER = "pager";
+	
+	// 登陆用户SESSION-KEY
+	public static final String SESSION_USER_KEY = "LOGINED_USER";
 }
