+++ src/main/resources/application.properties
@@ -43,8 +43,9 @@ lebang_api_host_oauth_out=http://test.4009515151.com/api/lebang/oauth/signout
 scopes=all
 #state=914dd9be-4964-46a2-9e5a-48b3fe179798 
 redirect_uri=http://127.0.0.1:8080/oauth/token
-redirect_uri_login=http://127.0.0.1:8080/login
+redirect_uri_login=http://127.0.0.1:8080/admin/login
 client_id=c9f3e1daecea7b3ee8c3204e73b27123
 response_type=code
 sign=y
 client_secret=7UcvvHD8jCZGx6oLeOYKumROgIn8SBpW
+server.error.path=error/error
